Renamed some minor things
