+++ packages/frontend/components/lobby/lobbyDesktop.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { Grid } from 'semantic-ui-react';
 
-import Layout from '../layout/Layout';
+import Layout from '../layout/layout';
 import GamesMenu from './rightPanel/gamesMenu';
 import HomeIndex from './leftPanel/homeIndex';
 import Chat from './mainPanel/chat';
